Add the ability to parse postgres headers at compile-time
This commit allows us to parse the postgres headers and generate
rust-bindings at compile-time, instead of relying on the pre-generated
bindings, with `--features=parse_headers`. This enables us to compile on
architectures for which we do not already have bindings generated, and
allows us to generate release artifacts using precisely the headers
found on the relevant machine.