Use well-known MemoryContexts instead of our own
I turns out that sigsetjmp is suprisingly expensive, 500ns in some of
our benchmarks, making it too expensive to call on every tuple of a
large dataset. Since we have no other way to clean up rust's global
allocator when postgres throws an exception, this means that our current
scheme of allocating in the TopTransactionContext unless switched to
another isn't really workable as a postgres exception can result in the
global allocator pointing at a deallocated memory context. Instead we'll
just use CurrentMemoryContext.